+++ test/test_column_string.cpp
@@ -845,6 +845,10 @@ TEST(ColumnString_FindAll_NoDuplicatesWithIndex)
     col.find_all(res, "a", 0, npos);
 
     CHECK_EQUAL(1, res.size());
+
+    // Clean-up
+    res.destroy();
+    col.destroy();
 }
 
 TEST(ColumnString_Count)
@@ -914,7 +918,7 @@ TEST(ColumnString_Index)
     asc.add("15");
     asc.add("HEJSA"); // 16
 
-    const StringIndex& ndx = asc.create_search_index();
+    const StringIndex& ndx = *asc.create_search_index();
     CHECK(asc.has_search_index());
 #ifdef TIGHTDB_DEBUG
     ndx.verify_entries(asc);